schema_registry/offset_waiter: remove offset_conflict values
This is fiddly to get right, because to ensure they see it,
writers would have to register their wait before issuing
their write, or we would have to keep some history for
the benefit of writers that wait() on an offset after
it has been replayed.

Instead, writers can detect conflicts independently
based on the offset returned from their write.

Signed-off-by: John Spray <jcs@vectorized.io>