config: add kafka_rpc_server_stream_recv_buf
Seastar's input_stream can buffer up to 128k
by default.  Sometimes we would rather it didn't,
e.g. if we have very large numbers of clients or
if we are trying to reduce queuing points for
latency.

This property is specific to the Kafka RPC server,
because Kafka client connections are the numerous
ones: internal RPC connections are usually smaller
in number and deserve to take large buffers if they
want to.