raft: refuse append_entries with data when disk critically full
...for non-controller partitions.

The consequence will be for too-full nodes to
refuse to take part in quorum writes and probably
result in loss of leadership.

This will leave the system in a degraded state, but
allow the controller group operable, so that it
can do things like deleting topics and/or changing
retention settings.