s/disk_log_impl: finish compaction before deleting segments
During compaction we loop through all of the segments until we find
first one which was compacted. When segments are deleted in
`disk_log_impl::remove()` we must make sure that compaction is already
finished to prevent iteration from accessing removed segment.

This problem caused segmentation fault that occurred while compacted
partition was moved from one of the nodes and finally removed.

Signed-off-by: Michal Maslanka <michal@vectorized.io>