redpanda: shutdown quota manager after kafka server is stopped
Quota manager refrence is a part of request context. We need to make
sure that quota manager is available as long as the kafka server is
running. Moved quota manager initialization before waiting for Kafka
server shutdown to make sure that its reference is available for all
requests that are being processed.