tests: per-node extra_node_conf in RedpandaService
This supports the use case of RackAwareReplicaPlacementTest,
without forcing the test to individually start each node
to pass the node config into it.

Instead, they can pass a template for node configuration
into the constructor (this is applied to all nodes), and/or
later set individual node's configuration with set_extra_node_conf.