tests/clients: add permissive mode for rpk describe_topic
When using this function to query leadership for partitions,
it is not necessary to exclude partitions just because
they failed to get some metadata from the leader (e.g. NOT_LEADER
errors for offets during transient leaderhsip change).

Add a `tolerant` flag that permits returning partially populated
RpkPartition results that just show the leader of a partition.