tests: improve logging on failure
It is helpful to print the error right at the point
of failure, rather than after the (potentially long
running) backtrace decode & log search jobs.

It'll get printed again later as well, but this way
I can search from the start of the file for the exception
name, and jump straight to the timestamp of the failure.