tests/clients: summary mode for describe group
The default mode is rather expensive for high partition counts,
and complicates handling systems in transient states when one
or more of the partitions is likely to be underoing leadership
movement and therefore have NOT_LEADER errors etc in the
default per-partition output.

When all we want to know is the group's state, this lets
us get that.