cluster: only delete cloud storage on partition deletion
Previously the delete_partition path did not distinguish
between actually deleting a partition (globally) and removing
the state for a partition on one node.

It is a little awkward to have a function with two modes
rather than two separate series of functions, but it
is desirable for the remote data deletion to happen
within partition_manager::remove, as it should happen
after local state is removed, but before we throw
away the `partition` object.