cluster/node: track data + cache drive separately
This class was previously holding a vector of disks, although
only populating it to length 1.  Instead, make it aware of
the distinction between data + cache drives, and handle
the case where they are the same drive.

In health monitor code where we previously iterated over drives,
we now use the data drive specifically.