cloud_storage: update remote_partition::erase
Now that we have the scrubber, it is not worth
trying to erase the partition from all peers: we
can try from one peer, and if anything goes wrong
with that, the scrubber will clean up later.

We don't strictly need to erase from remote_partition
at all any more, but it's a nicer behavior in the
happy path: the user sees the deletion work happen
at the time they delete the topic, rather than at
some indeterminate future time, and we will distribute
the deletion work on the same shards as the partitions
lived on.